Split view component, added unit test for drill title
