+++ frontend/src/metabase/modes/components/drill/UnderlyingRecordsDrill.unit.spec.js
@@ -1,24 +1,14 @@
 import { assocIn } from "icepick";
 import moment from "moment-timezone";
 import { ORDERS, PEOPLE } from "__support__/sample_database_fixture";
-
-import UnderlyingRecordsDrill from "metabase/modes/components/drill/UnderlyingRecordsDrill";
-
-function getActionProps(query, value) {
-  return {
-    question: query.question(),
-    clicked: {
-      column: query.aggregationDimensions()[0].column(),
-      value: 42,
-      dimensions: [
-        {
-          column: query.breakouts()[0].dimension().column(),
-          value: value,
-        },
-      ],
-    },
-  };
-}
+import { createMockState } from "metabase-types/store/mocks";
+import { createEntitiesState } from "__support__/store";
+import {
+  createSampleDatabase,
+  ORDERS_ID,
+} from "metabase-types/api/mocks/presets";
+import { getMetadata } from "metabase/selectors/metadata";
+import UnderlyingRecordsDrill from "./UnderlyingRecordsDrill";
 
 describe("UnderlyingRecordsDrill", () => {
   it("should not be valid for top level actions", () => {
@@ -26,6 +16,7 @@ describe("UnderlyingRecordsDrill", () => {
       UnderlyingRecordsDrill({ question: ORDERS.newQuestion() }),
     ).toHaveLength(0);
   });
+
   it("should be return correct new card for breakout by month", () => {
     const value = "2018-01-01T00:00:00Z";
     const query = ORDERS.query()
@@ -33,6 +24,7 @@ describe("UnderlyingRecordsDrill", () => {
       .breakout(["field", ORDERS.CREATED_AT.id, { "temporal-unit": "month" }]);
     const actions = UnderlyingRecordsDrill(getActionProps(query, value));
     expect(actions).toHaveLength(1);
+
     const q = actions[0].question();
     expect(q.query().query()).toEqual({
       "source-table": ORDERS.id,
@@ -44,6 +36,7 @@ describe("UnderlyingRecordsDrill", () => {
     });
     expect(q.display()).toEqual("table");
   });
+
   it("should be return correct new card for breakout by day-of-week", () => {
     const value = 4; // corresponds to Wednesday
     const query = ORDERS.query()
@@ -152,4 +145,79 @@ describe("UnderlyingRecordsDrill", () => {
     });
     expect(q.display()).toEqual("table");
   });
+
+  describe("title", () => {
+    it('should return "See these records" title for entities with title longer than 20 chars', () => {
+      const actions = UnderlyingRecordsDrill(setup("LongLongLongTableTitle"));
+      expect(actions).toHaveLength(1);
+
+      const [action] = actions;
+      if (!("title" in action)) {
+        throw new Error("Received unexpected action type");
+      }
+
+      expect(action.title).toEqual("See these records");
+    });
+
+    it("should contain entity title for entities shorter than 21 chars", () => {
+      const actions = UnderlyingRecordsDrill(setup("SomeTitle"));
+      expect(actions).toHaveLength(1);
+
+      const [action] = actions;
+      if (!("title" in action)) {
+        throw new Error("Received unexpected action type");
+      }
+
+      expect(action.title).toEqual("See these SomeTitles");
+    });
+  });
 });
+
+function getActionProps(query, value) {
+  return {
+    question: query.question(),
+    clicked: {
+      column: query.aggregationDimensions()[0].column(),
+      value: 42,
+      dimensions: [
+        {
+          column: query.breakouts()[0].dimension().column(),
+          value: value,
+        },
+      ],
+    },
+  };
+}
+
+function getMockTable(tableDisplayName) {
+  const state = createMockState({
+    entities: createEntitiesState({
+      databases: [createSampleDatabase()],
+    }),
+  });
+
+  const metadata = getMetadata(state);
+  const table = metadata.table(ORDERS_ID);
+
+  table.display_name = tableDisplayName;
+
+  return table;
+}
+
+function setup(tableDisplayName) {
+  const table = getMockTable(tableDisplayName);
+
+  return {
+    question: table.newQuestion(),
+    clicked: {
+      column: table.fields[0].column(),
+      value: 42,
+      dimensions: [
+        {
+          column: table.fields[0].column(),
+          value: 42,
+        },
+      ],
+    },
+  };
+}
