+++ frontend/src/metabase/visualizations/components/ChartClickActions/ChartClickActions.css
